+++ src/cirrus/git_tools.py
@@ -5,7 +5,13 @@ _git_tools_
 Utils for doing git and github related business
 
 """
+import os
 import git
+
+from cirrus.logger import get_logger
+
+
+LOGGER = get_logger()
 
 
 def checkout_and_pull(repo_dir, branch_from):
@@ -47,11 +53,60 @@ def branch(repo_dir, branchname, branch_from):
         g.checkout(branch_from, b=branchname)
 
     if not str(repo.active_branch) == branchname:
-        msg = "uh oh, not on expected branch"
-        raise RuntimeError((
-            "{0}. You are here -> {1}, "
-            "you expected to be here -> {2}"
-            ).format(msg, repo.active_branch, branchname))
+        msg = (
+            "Error: Not able to checkout expected branch"
+            "You are here -> {0}, "
+            "you expected to be here -> {1}"
+            ).format(repo.active_branch, branchname)
+        LOGGER.error(msg)
+        raise RuntimeError(msg)
+
+
+def update_to_branch(branch, config, origin='origin'):
+    """
+    checkout specified branch, updating to pull in latest remotes
+
+    """
+    LOGGER.info(
+        "selfupdate running, will switch to branch {0}".format(
+            branch
+        )
+    )
+    repo_dir = os.getcwd()
+
+    LOGGER.info("fetching remotes...")
+    r = git.Repo(repo_dir)
+    r.remotes[origin].fetch()
+
+    g = git.Git()
+    LOGGER.info("checking out {0}...".format(branch))
+    g.checkout('{0}/{1}'.format(origin, branch), b=branch)
+
+    branch_ref = r.heads[branch]
+    branch_ref.checkout()
+    return
+
+
+def update_to_tag(tag, config, origin='origin'):
+    """
+    checkout specified tag, pulling remote tags first
+    """
+    LOGGER.info(
+        "selfupdate running, will switch to tag {0}".format(
+            tag
+        )
+    )
+    repo_dir = os.getcwd()
+
+    LOGGER.info("fetching remote tags...")
+    r = git.Repo(repo_dir)
+    r.remotes[origin].fetch(tags=True)
+
+    ref = r.tags[tag]
+    LOGGER.info("checking out {0}...".format(tag))
+    g = git.Git()
+    g.checkout(ref)
+    return
 
 
 def commit_files(repo_dir, commit_msg, *filenames):
